% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Custom Radar Coordinate System}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{Variable to map angle to ('x' or 'y').}

\item{start}{Offset of starting point from 12 o'clock in radians.}

\item{direction}{1, clockwise; -1, anticlockwise.}
}
\description{
Custom Radar Coordinate System
}
\keyword{internal}
