% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-radar.R
\name{is.linear.radar}
\alias{is.linear.radar}
\title{Check if coordinate system is linear (for radar plots)}
\usage{
is.linear.radar(coord)
}
\arguments{
\item{coord}{Coordinate system object.}
}
\description{
Check if coordinate system is linear (for radar plots)
}
\keyword{internal}
