% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_star_plot.R
\name{plot_pcoa_star}
\alias{plot_pcoa_star}
\title{Plot a Star Plot of Mean PCoA Position for Groups}
\usage{
plot_pcoa_star(
  physeq,
  sample_var,
  colors_all,
  view_type = "together",
  error_bar = "IQR",
  fill_alpha = 0.4,
  distance
)
}
\arguments{
\item{physeq}{A phyloseq object containing your microbiome data.}

\item{sample_var}{A character string. The name of the column in your sample data
to use for grouping samples (e.g., "location", "treatment").}

\item{view_type}{A character string, either "together" (default) or "separate".
Determines whether groups are plotted on the same plot or faceted.}

\item{error_bar}{A character string, one of "IQR" (default), "SE", or "none".
Determines the type of error bars to display.}

\item{fill_alpha}{A numeric value between 0 and 1 (default 0.4).
Controls the transparency of the polygon fill under the star plot.}

\item{distance}{An accepted phyloseq ordination type ("bray" for example)}
}
\value{
A ggplot object representing the star plot of PCoA centroids.
}
\description{
This function calculates dissimilarity, performs PCoA, and plots the
mean position of sample groups along the first 5 PCoA axes using a star (radar) plot.
}
\examples{
# ...
#   plot_pcoa_star(
#     physeq = gp_subset,
#     sample_var = "SampleType",
#     colors_all = c(...),
#     view_type = "together",
#     error_bar = "SE",
#     fill_alpha = 0.2,
#     distance = "bray"
#    )
#
}
