% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_star_plot.R
\name{plot_taxa_star}
\alias{plot_taxa_star}
\title{Plot a Star Plot of Relative Abundance for Three Taxa}
\usage{
plot_taxa_star(physeq, sample_var, taxa_rank, taxa_names)
}
\arguments{
\item{physeq}{A phyloseq object containing your microbiome data.}

\item{sample_var}{A character string. The name of the column in your sample data
to use for faceting (e.g., "location", "treatment").}

\item{taxa_rank}{A character string. The taxonomic rank you are providing taxa
names for (e.g., "Phylum", "Genus").}

\item{taxa_names}{A character vector containing the names of the three specific
taxa you want to plot.}
}
\value{
A ggplot object representing the star plot.
}
\description{
This function generates a star (radar) plot showing the relative abundance
of three specified taxa plus an "Other" category, faceted by a sample variable.
}
\examples{
# This example won't run unless you have the 'GlobalPatterns' dataset loaded
# if (requireNamespace("phyloseq", quietly = TRUE)) {
#   data(GlobalPatterns)
#
#   # Prune to a manageable size for example
#   gp_subset <- prune_samples(sample_sums(GlobalPatterns) > 500000, GlobalPatterns)
#   gp_subset <- prune_taxa(taxa_sums(gp_subset) > 0, gp_subset)
#
#   plot_taxa_star(
#     physeq = gp_subset,
#     sample_var = "SampleType",
#     taxa_rank = "Phylum",
#     taxa_names = c("Firmicutes", "Bacteroidetes", "Proteobacteria")
#    )
# }
}
